Fix the other MITS 500 error

+++ uber/site_sections/mits.py
@@ -276,7 +276,8 @@ class Root:
                         session.delete(game.guidebook_header)
                     session.add(header_pic)
                 if thumbnail_pic:
-                    session.delete(game.guidebook_thumbnail)
+                    if game.guidebook_thumbnail:
+                        session.delete(game.guidebook_thumbnail)
                     session.add(thumbnail_pic)
                 session.add(game)
                 raise HTTPRedirect('index?message={}', 'Game saved')
